requirements.txt
Flask>=3.0.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10
dashscope>=1.14.0
